Improvements and Progress
Finished the Walls class. Added game states, and game modes (just the
enums... implementations have not been completed. Added wall drawing to
Draw, and placed background drawing in Draw where it belongs, instead of
keeping it directly in Game. Began the hash set of pressed keys, to
allow for the registering of multiple key strokes at once. This will
allow for local multiplayer with one player using wasd, and the other
using the arrow keys.