Nasty Bugs
There are currently some strange bugs with the menus:

1. The Main Menu is loaded first EVERY TIME the game is started,
however, it sometimes doesn't display properly. The reason is unknown,
and it seems to occur at random. Fresh starts either load a blank JPanel
as the Main Menu, load in the correct black JPanel with the neon border
but without all the buttons and labels, or it loads up properly.
Possible issue with unintended persistence? I don't really know.

2. When attempting to swap JPanels from the main menu to the game
renderer, the window freezes at the Gaps Options view, with one of the
buttons depressed. The game never displays. The reason is unknown, but I
have a feeling it has something to do with my setState() method never
returning, locking the MainMenu into a "waiting" state and then
attempting to delete the still running menu.

I will try to fix this at a later date.