Fixed Graphics and Added Snake AI
The game is drawn properly once again, and the pause/gameover menus
layer on top of the game (providing a less jarring effect when pausing
and unpausing).
Simplified the main menu to be one page, while having gaps and fruit
options menus for the remaining three game choice pages.
Added a simple snake AI that runs when the game mode is SOLO_VS. It
works well for the most part, but has some cases where it makes
unnecessary decisions that get it killed. Occasionally the snake fails
to turn when it has the ability to turn and could prolong its life if it
turned. The reason is unknown. The AI is good enough for our purposes,
and if it was perfect, you would almost never win, so I will likely
leave it as is.
The next step will be to add the High Scores and persistence, as well as
a placeholder page explaining that multiplayer has not yet been
implemented (I am still researching how to make it work. UDP with some
layers making out of date messages be ignored is what it looks like I'll
need. I'll likely set up some AWS to provide a list of peers, and
perform NAT traversal).