Added High Scores
Added functionality for keeping track of high scores in solo and player
vs computer games. There is a bug in the implementation causing a game
end which should open the high scores panel to instead open the pause
panel. A quick look didn't let me find the issue, so a fix will have to
wait for the moment.