High Scores Panel Fix
The game now properly loads the high scores insertion panel when a top 5
score is achieved in solo or solo_vs play.
Other bugs are present.
1. Persistence isn't working. Scores are not being saved.
2. Since my PC has been plagued by blur issues after an update, I had to
disable auto settings for high DPI displays. After doing this for
eclipse, the game now loads with a gap under the bottom wall. I don't
know why, but it doesn't really affect the game, so it is a low priority
bug.
3. I believe when I clicked submit once on the high scores panel, the
code should've rejected my input, but it didn't. Look into that.