Fixed Persistence
The game now properly saves high scores. The bug was a < where there
should have been a > in the compareTo method of Score. A unit test
narrowed it down.
There is now a bug with the high scores input menu. When the score is
below the top 5, the menu is still displayed, and one of the text boxes
is modifiable. The submit button also seems to fail in this case. The
menu should only be displayed when a top 5 score is achieved.